# Project/1.PondPlanning/docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI-based Village Pond Planning System",
    "description": "Terrain, catchment and runoff analysis for village pond siting. Analysis routes are asynchronous: they return 202 with a job id to poll.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/terrain/{village_id}/layers": {
      "get": {
        "tags": [
          "terrain"
        ],
        "summary": "List Layers",
        "description": "Every toggleable layer for this village \u2014 the source of FR8's overlay list.",
        "operationId": "list_layers_api_v1_terrain__village_id__layers_get",
        "parameters": [
          {
            "name": "village_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Village identifier",
              "title": "Village Id"
            },
            "description": "Village identifier"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TerrainLayers"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/terrain/{village_id}/dem": {
      "get": {
        "tags": [
          "terrain"
        ],
        "summary": "Get Dem Asset",
        "description": "Provenance and accuracy of the elevation model underneath every terrain figure.\n\nA first-class endpoint rather than a footnote: the credibility of the storage\nestimate depends on the reader knowing the source is 30 m.",
        "operationId": "get_dem_asset_api_v1_terrain__village_id__dem_get",
        "parameters": [
          {
            "name": "village_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Village identifier",
              "title": "Village Id"
            },
            "description": "Village identifier"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DEMAsset"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/terrain/{village_id}/contours": {
      "get": {
        "tags": [
          "terrain"
        ],
        "summary": "Get Contours",
        "description": "FR2: contours generated from the DEM pipeline at the requested interval.",
        "operationId": "get_contours_api_v1_terrain__village_id__contours_get",
        "parameters": [
          {
            "name": "village_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Village identifier",
              "title": "Village Id"
            },
            "description": "Village identifier"
          },
          {
            "name": "interval",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 50,
              "exclusiveMinimum": 0,
              "description": "Contour interval in metres",
              "default": 2.0,
              "title": "Interval"
            },
            "description": "Contour interval in metres"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContourResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/terrain/{village_id}/streams": {
      "get": {
        "tags": [
          "terrain"
        ],
        "summary": "Get Streams",
        "description": "The modelled drainage network, with the threshold that produced it.",
        "operationId": "get_streams_api_v1_terrain__village_id__streams_get",
        "parameters": [
          {
            "name": "village_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Village identifier",
              "title": "Village Id"
            },
            "description": "Village identifier"
          },
          {
            "name": "threshold",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "exclusiveMinimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Flow-accumulation cells",
              "title": "Threshold"
            },
            "description": "Flow-accumulation cells"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StreamNetwork"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/terrain/{village_id}/derived/{surface}": {
      "get": {
        "tags": [
          "terrain"
        ],
        "summary": "Get Derived Surface",
        "description": "Slope, aspect, curvature, TWI, hillshade or flow accumulation.",
        "operationId": "get_derived_surface_api_v1_terrain__village_id__derived__surface__get",
        "parameters": [
          {
            "name": "village_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Village identifier",
              "title": "Village Id"
            },
            "description": "Village identifier"
          },
          {
            "name": "surface",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "slope",
                "aspect",
                "curvature",
                "twi",
                "hillshade",
                "flow_accumulation"
              ],
              "type": "string",
              "description": "Which surface to return",
              "title": "Surface"
            },
            "description": "Which surface to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DerivedSurfaceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ContourResponse": {
        "properties": {
          "village_id": {
            "type": "string",
            "format": "uuid",
            "title": "Village Id"
          },
          "interval": {
            "$ref": "#/components/schemas/QuantityOut"
          },
          "levels": {
            "type": "integer",
            "title": "Levels",
            "description": "Number of distinct elevation levels returned"
          },
          "vertices_before_simplification": {
            "type": "integer",
            "title": "Vertices Before Simplification"
          },
          "vertices_after_simplification": {
            "type": "integer",
            "title": "Vertices After Simplification"
          },
          "simplification_tolerance": {
            "$ref": "#/components/schemas/QuantityOut"
          },
          "geojson": {
            "$ref": "#/components/schemas/GeoJSONFeatureCollection"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/ResultWarning"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "village_id",
          "interval",
          "levels",
          "vertices_before_simplification",
          "vertices_after_simplification",
          "simplification_tolerance",
          "geojson"
        ],
        "title": "ContourResponse",
        "description": "FR2: contours generated from the DEM pipeline at a requested interval."
      },
...
      "DEMAsset": {
        "properties": {
          "village_id": {
            "type": "string",
            "format": "uuid",
            "title": "Village Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "native_resolution": {
            "$ref": "#/components/schemas/QuantityOut"
          },
          "working_resolution": {
            "$ref": "#/components/schemas/QuantityOut"
          },
          "vertical_accuracy_relative": {
            "$ref": "#/components/schemas/QuantityOut"
          },
          "vertical_accuracy_absolute": {
            "$ref": "#/components/schemas/QuantityOut"
          },
          "crs": {
            "type": "string",
            "title": "Crs"
          },
          "acquired": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acquired"
          },
          "attribution": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Attribution"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/ResultWarning"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "village_id",
          "source",
          "native_resolution",
          "working_resolution",
          "vertical_accuracy_relative",
          "vertical_accuracy_absolute",
          "crs",
          "attribution"
        ],
        "title": "DEMAsset",
        "description": "Provenance of the elevation model every terrain number rests on.\n\nExposed as an endpoint rather than buried in a log because the honesty of\nevery downstream figure depends on it: a 1 m contour interval interpolated\nfrom a 30 m source is interpolated precision, not measured accuracy."
      },
...
      "DerivedSurfaceResponse": {
        "properties": {
          "village_id": {
            "type": "string",
            "format": "uuid",
            "title": "Village Id"
          },
          "surface": {
            "type": "string",
            "enum": [
              "slope",
              "aspect",
              "curvature",
              "twi",
              "hillshade",
              "flow_accumulation"
            ],
            "title": "Surface"
          },
          "algorithm": {
            "type": "string",
            "title": "Algorithm",
            "description": "e.g. 'Horn (1981) 3x3 finite difference'"
          },
          "statistics": {
            "additionalProperties": {
              "$ref": "#/components/schemas/QuantityOut"
            },
            "type": "object",
            "title": "Statistics"
          },
          "layer": {
            "$ref": "#/components/schemas/LayerDescriptor"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/ResultWarning"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "village_id",
          "surface",
          "algorithm",
          "statistics",
          "layer"
        ],
        "title": "DerivedSurfaceResponse",
        "description": "Slope, aspect, curvature, TWI, hillshade or flow accumulation as a layer."
      },
...
      "GeoJSONFeature": {
        "properties": {
          "type": {
            "type": "string",
            "const": "Feature",
            "title": "Type",
            "default": "Feature"
          },
          "geometry": {
            "additionalProperties": true,
            "type": "object",
            "title": "Geometry"
          },
          "properties": {
            "additionalProperties": true,
            "type": "object",
            "title": "Properties"
          }
        },
        "type": "object",
        "required": [
          "geometry"
        ],
        "title": "GeoJSONFeature",
        "description": "A GeoJSON Feature, loosely typed.\n\nGeometry is not modelled field-by-field on purpose: the map client consumes\nit directly, and a hand-written GeoJSON type is a large surface that adds no\nsafety the client actually uses. Geometry validity is enforced in PostGIS and\nin the engines, where it matters."
      },
      "GeoJSONFeatureCollection": {
        "properties": {
          "type": {
            "type": "string",
            "const": "FeatureCollection",
            "title": "Type",
            "default": "FeatureCollection"
          },
          "features": {
            "items": {
              "$ref": "#/components/schemas/GeoJSONFeature"
            },
            "type": "array",
            "title": "Features"
          },
          "crs": {
            "type": "string",
            "title": "Crs",
            "description": "Always EPSG:4326 on the wire",
            "default": "EPSG:4326"
          }
        },
        "type": "object",
        "required": [
          "features"
        ],
        "title": "GeoJSONFeatureCollection",
        "description": "A GeoJSON FeatureCollection, always in EPSG:4326 on the wire.\n\nEvery computation happens in the UTM zone derived from the data's own\ncentroid; 4326 is the interchange format only. ``crs`` is stated explicitly\nso no consumer has to assume."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LayerDescriptor": {
        "properties": {
          "layer_id": {
            "type": "string",
            "title": "Layer Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "raster",
              "vector"
            ],
            "title": "Kind"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "tile_url_template": {
            "type": "string",
            "title": "Tile Url Template"
          },
          "legend_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Legend Url"
          },
          "units": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Units"
          },
          "value_range": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value Range",
            "description": "[min, max] for the legend"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "description": "What this layer was derived from"
          }
        },
        "type": "object",
        "required": [
          "layer_id",
          "kind",
          "title",
          "tile_url_template",
          "source"
        ],
        "title": "LayerDescriptor",
        "description": "Everything the map client needs to add one layer and label it honestly."
      },
...
      "QuantityOut": {
        "properties": {
          "value": {
            "type": "number",
            "title": "Value"
          },
          "unit": {
            "$ref": "#/components/schemas/Unit"
          },
          "uncertainty_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uncertainty Pct",
            "description": "Symmetric relative uncertainty in percent. null means exact by construction."
          },
          "low": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Low",
            "description": "Lower bound of the uncertainty band"
          },
          "high": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "High",
            "description": "Upper bound of the uncertainty band"
          },
          "method": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Method",
            "description": "How the value was derived"
          },
          "display": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display",
            "description": "Preformatted for direct rendering"
          }
        },
        "type": "object",
        "required": [
          "value",
          "unit"
        ],
        "title": "QuantityOut",
        "description": "A number with its unit, uncertainty band and provenance.\n\nThe wire counterpart of :class:`app.domain.units.Quantity`. ``display`` is\nincluded deliberately: it removes any chance of the frontend rendering\n\"18950\" because someone forgot to append the unit, and it is what appears in\nreport screenshots.",
        "example": {
          "display": "18,950.00 m\u00b3 (\u00b120 %)",
          "high": 22740.0,
          "low": 15160.0,
          "method": "EAV curve integrated over the filled depression",
          "uncertainty_pct": 20.0,
          "unit": "m3",
          "value": 18950.0
        }
      },
...
      "ResultWarning": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "severity": {
            "type": "string",
            "enum": [
              "info",
              "caution",
              "critical"
            ],
            "title": "Severity",
            "default": "caution"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ResultWarning",
        "description": "A caveat attached to an otherwise successful result.\n\nWarnings are how this system avoids presenting a number more confidently than\nit deserves \u2014 \"the source DEM is 30 m, so relief below ~5 m is interpolated,\nnot measured\" belongs next to the storage estimate, not buried in a report."
      },
...
      "StreamNetwork": {
        "properties": {
          "village_id": {
            "type": "string",
            "format": "uuid",
            "title": "Village Id"
          },
          "accumulation_threshold": {
            "$ref": "#/components/schemas/QuantityOut"
          },
          "total_length": {
            "$ref": "#/components/schemas/QuantityOut"
          },
          "strahler_max_order": {
            "type": "integer",
            "title": "Strahler Max Order"
          },
          "geojson": {
            "$ref": "#/components/schemas/GeoJSONFeatureCollection"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/ResultWarning"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "village_id",
          "accumulation_threshold",
          "total_length",
          "strahler_max_order",
          "geojson"
        ],
        "title": "StreamNetwork",
        "description": "Extracted drainage network, with the threshold that produced it.\n\n``accumulation_threshold`` is surfaced because it is the one knob that\nvisibly changes the answer: too high and tributaries vanish, too low and the\nhillsides fill with spurious channels."
      },
...
      "TerrainLayers": {
        "properties": {
          "village_id": {
            "type": "string",
            "format": "uuid",
            "title": "Village Id"
          },
          "layers": {
            "items": {
              "$ref": "#/components/schemas/LayerDescriptor"
            },
            "type": "array",
            "title": "Layers"
          }
        },
        "type": "object",
        "required": [
          "village_id",
          "layers"
        ],
        "title": "TerrainLayers",
        "description": "The full toggleable layer set for a village \u2014 FR8's six overlays live here."
      },
      "Unit": {
        "type": "string",
        "enum": [
          "m",
          "m2",
          "ha",
          "km2",
          "m3",
          "mm",
          "mm/yr",
          "deg",
          "%",
          "ratio",
          "count",
          "yr",
          "d",
          "INR"
        ],
        "title": "Unit",
        "description": "Units used anywhere in the system. SI, except where the domain differs.\n\nIndian minor-irrigation practice reports catchment in hectares and storage in\ncubic metres, so those are the reporting units even though the engine\ncomputes in square metres."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}